+++ pybillomat/recurring_items.py
@@ -8,144 +8,82 @@ Recurring-Items
 """
 
 
+import errors
+import xml.etree.ElementTree as ET
 from bunch import Bunch
 from http import Url
-from _tools import Item, ItemsIterator
+from _items_base import Item, ItemsIterator
 
 
-# def _recurring_xml(
-#     client_id = None,
-#     contact_id = None,
-#     title = None,
-#     address = None,
-#     supply_date = None,
-#     supply_date_type = None,
-#     due_days = None,
-#     discount_rate = None,
-#     discount_days = None,
-#     name = None,
-#     label = None,
-#     intro = None,
-#     note = None,
-#     currency_code = None,
-#     reduction = None,
-#     net_gross = None,
-#     quote = None,
-#     payment_types = None,
-#     action = None,
-#     cycle_number = None,
-#     cycle = None,
-#     hour = None,
-#     start_date = None,
-#     end_date = None,
-#     next_creation_date = None,
-#     email_sender = None,
-#     email_subject = None,
-#     email_message = None,
-#     email_filename = None,
-#     offer_id = None,
-#     confirmation_id = None,
-#     template_id = None,
-# ):
-#     """
-#     Creates the XML to add or edit a recurring
-#     """
-#
-#     integer_field_names = [
-#         "client_id",
-#         "contact_id",
-#         "due_days",
-#         "discount_rate",
-#         "discount_days",
-#         "cycle_number",
-#         "hour",
-#         "offer_id",
-#         "confirmation_id",
-#         "template_id",
-#     ]
-#     date_or_string_fieldnames = [
-#         "supply_date",
-#         "next_creation_date",
-#     ]
-#     date_fieldnames = [
-#         "start_date",
-#         "end_date",
-#     ]
-#     float_fieldnames = [
-#         "quote",
-#     ]
-#     string_fieldnames = [
-#         "title",
-#         "address",
-#         "supply_date_type",
-#         "name",
-#         "label",
-#         "intro",
-#         "note",
-#         "currency_code",
-#         "reduction",
-#         "net_gross",
-#         "payment_types",
-#         "action",
-#         "cycle",
-#         "email_sender",
-#         "email_subject",
-#         "email_message",
-#         "email_filename",
-#     ]
-#
-#     recurring_tag = ET.Element("recurring")
-#
-#     # Integer Fields
-#     for field_name in integer_field_names:
-#         value = locals()[field_name]
-#         if value is not None:
-#             new_tag = ET.Element(field_name)
-#             new_tag.text = unicode(int(value))
-#             recurring_tag.append(new_tag)
-#
-#     # Date or string fields
-#     for field_name in date_or_string_fieldnames:
-#         value = locals()[field_name]
-#         if value is not None:
-#             new_tag = ET.Element(field_name)
-#             if isinstance(value, datetime.date):
-#                 new_tag.text = value.isoformat()
-#             else:
-#                 new_tag.text = unicode(value)
-#             recurring_tag.append(new_tag)
-#
-#     # Date fields
-#     for field_name in date_fieldnames:
-#         value = locals()[field_name]
-#         if value is not None:
-#             new_tag = ET.Element(field_name)
-#             new_tag.text = value.isoformat()
-#             recurring_tag.append(new_tag)
-#
-#     # Float Fields
-#     for field_name in float_fieldnames:
-#         value = locals()[field_name]
-#         if value is not None:
-#             new_tag = ET.Element(field_name)
-#             new_tag.text = unicode(float(value))
-#             recurring_tag.append(new_tag)
-#
-#     # String Fields
-#     for field_name in string_fieldnames:
-#         value = locals()[field_name]
-#         if value is not None:
-#             new_tag = ET.Element(field_name)
-#             new_tag.text = unicode(value)
-#             recurring_tag.append(new_tag)
-#
-#     xml = ET.tostring(recurring_tag)
-#
-#     # Finished
-#     return xml
+def _recurring_item_xml(
+    recurring_id = None,
+    article_id = None,
+    unit = None,
+    quantity = None,
+    unit_price = None,
+    tax_name = None,
+    tax_rate = None,
+    title = None,
+    description = None,
+    reduction = None
+):
+    """
+    Creates the XML to add or edit a recurring-item
+    """
+
+    integer_field_names = [
+        "recurring_id",
+        "article_id",
+    ]
+    float_fieldnames = [
+        "quantity",
+        "unit_price",
+        "tax_rate",
+    ]
+    string_fieldnames = [
+        "unit",
+        "tax_name",
+        "title",
+        "description",
+        "reduction",
+    ]
+
+    recurring_item_tag = ET.Element("recurring-item")
+
+    # Integer Fields
+    for field_name in integer_field_names:
+        value = locals()[field_name]
+        if value is not None:
+            new_tag = ET.Element(field_name)
+            new_tag.text = unicode(int(value))
+            recurring_item_tag.append(new_tag)
+
+    # Float Fields
+    for field_name in float_fieldnames:
+        value = locals()[field_name]
+        if value is not None:
+            new_tag = ET.Element(field_name)
+            new_tag.text = unicode(float(value))
+            recurring_item_tag.append(new_tag)
+
+    # String Fields
+    for field_name in string_fieldnames:
+        value = locals()[field_name]
+        if value is not None:
+            new_tag = ET.Element(field_name)
+            new_tag.text = unicode(value)
+            recurring_item_tag.append(new_tag)
+
+    xml = ET.tostring(recurring_item_tag)
+
+    # Finished
+    return xml
 
 
 class RecurringItem(Item):
+
+    base_path = u"/api/recurring-items"
+
 
     def __init__(self, conn, id = None, recurring_item_etree = None):
         """
@@ -157,7 +95,6 @@ class RecurringItem(Item):
         Bunch.__init__(self)
 
         self.conn = conn
-        self.base_path = u"/api/recurring-items"
 
         self.id = id  # integer
         self.created = None  # datetime
@@ -173,174 +110,75 @@ class RecurringItem(Item):
         self.description = None
         self.total_gross = None  # float
         self.total_net = None  # float
+        self.reduction = None
+        self.total_gross_unreduced = None  # float
+        self.total_net_unreduced = None  # float
 
         if recurring_item_etree is not None:
             self.load_from_etree(recurring_item_etree)
 
 
-#     @classmethod
-#     def create(
-#         cls,
-#         conn,
-#         client_id = None,
-#         contact_id = None,
-#         title = None,
-#         address = None,
-#         supply_date = None,
-#         supply_date_type = None,
-#         due_days = None,
-#         discount_rate = None,
-#         discount_days = None,
-#         name = None,
-#         label = None,
-#         intro = None,
-#         note = None,
-#         currency_code = None,
-#         reduction = None,
-#         net_gross = None,
-#         quote = None,
-#         payment_types = None,
-#         action = None,
-#         cycle_number = None,
-#         cycle = None,
-#         hour = None,
-#         start_date = None,
-#         end_date = None,
-#         next_creation_date = None,
-#         email_sender = None,
-#         email_subject = None,
-#         email_message = None,
-#         email_filename = None,
-#         offer_id = None,
-#         confirmation_id = None,
-#         template_id = None,
-#         recurring_items = None
-#     ):
-#         """
-#         Creates a recurring
-#
-#         :param conn: Connection-Object
-#
-#         :param client_id: ID of the client
-#         :param contact_id: ID of the contact
-#         :param title: Document title; Let it empty to use the default value
-#             from settings: "Invoice [Invoice.invoice_number]"
-#         :param address: the address;
-#             Pass an empty value to use the current customer address.
-#         :param supply_date: supply/delivery date; MIXED (DATE/ALNUM)
-#         :param supply_date_type: type or supply/delivery date; ALNUM (
-#             "SUPPLY_DATE", "DELIVERY_DATE", "SUPPLY_TEXT", "DELIVERY_TEXT")
-#         :param due_days: Due days
-#         :param discount_rate: Cash discount
-#         :param discount_days: Cash discount days
-#         :param name: Name of the recurring; is the title of the recurring
-#         :param label: Label text to describe the project
-#         :param intro: Introductory text; Default value taken from the settings
-#         :param note: Explanatory notes; Default value taken from the settings
-#         :param reduction: Reduction (absolute or percent: 10/10%)
-#         :param currency_code: Currency; ISO currency code
-#         :param net_gross: Price basis (gross or net prices)
-#         :param quote: Currency quote (for conversion into standard currency)
-#         :param payment_types: List (separated by comma) of all accepted payment
-#             types.
-#         :param action: Action to be executed (CREATE, COMPLETE, EMAIL)
-#         :param cycle_number: Number of intervals. For example, 3 for
-#             "every 3 months"
-#         :param cycle: Interval (DAILY, WEEKLY, MONTHLY, YEARLY)
-#         :param hour: Time of Day (hour)
-#         :param start_date: Start date;
-#         :param end_date: End date
-#         :param next_creation_date: Date of the next creation;
-#             Put "" (empty string) to set recurring inactive.
-#         :param email_sender: Sender when sending e-mail. If you pass an empty
-#             value, the sender will be used from the settings.
-#         :param email_subject: Subject when sending e-mail. If you pass an
-#             empty value, the subject will be used from the settings.
-#         :param email_message: Message text when sending e-mail. If you pass
-#             an empty value, the message will be used from the settings.
-#         :param email_filename: Filename of the invoice when sending e-mail.
-#             If you pass an empty value, the filename will be used from the settings.
-#         :param offer_id: The ID of the estimate, if the recurring was created
-#             from an estimate.
-#         :param confirmation_id: The ID of the confirmation, if the recurring
-#             was created from a confirmation.
-#         :param template_id: Template ID
-#         :param recurring_items: List with RecurringItem-Objects
-#         """
-#
-#         # XML
-#         xml = _recurring_xml(
-#             client_id = client_id,
-#             contact_id = contact_id,
-#             title = title,
-#             address = address,
-#             supply_date = supply_date,
-#             supply_date_type = supply_date_type,
-#             due_days = due_days,
-#             discount_rate = discount_rate,
-#             discount_days = discount_days,
-#             name = name,
-#             label = label,
-#             intro = intro,
-#             note = note,
-#             currency_code = currency_code,
-#             reduction = reduction,
-#             net_gross = net_gross,
-#             quote = quote,
-#             payment_types = payment_types,
-#             action = action,
-#             cycle_number = cycle_number,
-#             cycle = cycle,
-#             hour = hour,
-#             start_date = start_date,
-#             end_date = end_date,
-#             next_creation_date = next_creation_date,
-#             email_sender = email_sender,
-#             email_subject = email_subject,
-#             email_message = email_message,
-#             email_filename = email_filename,
-#             offer_id = offer_id,
-#             confirmation_id = confirmation_id,
-#             template_id = template_id
-#         )
-#
-#         # Path
-#         path = "/api/recurrings"
-#
-#         # Send POST-request
-#         response = conn.post(path = path, body = xml)
-#         if response.status != 201:  # Created
-#             raise errors.BillomatError(unicode(response.data, encoding = "utf-8"))
-#
-#         # Create Recurring-Object
-#         recurring = cls(conn = conn)
-#         recurring.content_language = response.headers.get("content-language", None)
-#         recurring.load_from_xml(response.data)
-#
-#         # Finished
-#         return recurring
-#
-#
-#     def delete(self, id = None):
-#         """
-#         Deletes a recurring
-#         """
-#
-#         # Parameters
-#         if id:
-#             self.id = id
-#         if not self.id:
-#             raise errors.NoIdError()
-#
-#         # Path
-#         path = "/api/recurrings/{id}".format(id = self.id)
-#
-#         # Fetch data
-#         response = self.conn.delete(path = path)
-#         if response.status != 200:
-#             raise errors.BillomatError(unicode(response.data, encoding = "utf-8"))
-#
-#
+    @classmethod
+    def create(
+        cls,
+        conn,
+        recurring_id = None,
+        article_id = None,
+        unit = None,
+        quantity = None,
+        unit_price = None,
+        tax_name = None,
+        tax_rate = None,
+        title = None,
+        description = None,
+        reduction = None
+    ):
+        """
+        Creates a recurring-item
+
+        :param conn: Connection-Object
+
+        :param recurring_id: ID of the recurring
+        :param article_id: ID of the article, sets additionally the values
+            from the article on creation
+        :param unit: Unit
+        :param quantity: Quantity
+        :param unit_price: Price per unit
+        :param tax_name: Name of the tax
+        :param tax_rate: rate of taxation
+        :param title: Title
+        :param description: Description
+        :param reduction: Reduction (absolute or percent: 10/10%)
+        """
+
+        # XML
+        xml = _recurring_item_xml(
+            recurring_id = recurring_id,
+            article_id = article_id,
+            unit = unit,
+            quantity = quantity,
+            unit_price = unit_price,
+            tax_name = tax_name,
+            tax_rate = tax_rate,
+            title = title,
+            description = description,
+            reduction = reduction
+        )
+
+        # Send POST-request
+        response = conn.post(path = cls.base_path, body = xml)
+        if response.status != 201:  # Created
+            raise errors.BillomatError(unicode(response.data, encoding = "utf-8"))
+
+        # Create Recurring-Object
+        recurring = cls(conn = conn)
+        recurring.content_language = response.headers.get("content-language", None)
+        recurring.load_from_xml(response.data)
+
+        # Finished
+        return recurring
+
+
 #     def edit(
 #         self,
 #         id = None,