+++ pybillomat/invoices.py
@@ -12,10 +12,13 @@ import xml.etree.ElementTree as ET
 from bunch import Bunch
 from http import Url
 import errors
-from _items_base import ItemsIterator
+from _items_base import Item, ItemsIterator
 
 
-class Invoice(Bunch):
+class Invoice(Item):
+
+    base_path = u"/api/invoices"
+
 
     def __init__(self, conn, id = None, invoice_etree = None):
         """
@@ -87,74 +90,6 @@ class Invoice(Bunch):
             self.load_from_etree(invoice_etree)
 
 
-    def load_from_etree(self, etree_element):
-        """
-        Loads data from Element-Tree
-        """
-
-        for item in etree_element:
-
-            # Get data
-            isinstance(item, ET.Element)
-            tag = item.tag
-            type = item.attrib.get("type")
-            text = item.text
-
-            if text is not None:
-                if type == "integer":
-                    setattr(self, tag, int(text))
-                elif type == "datetime":
-                    # <created type="datetime">2011-10-04T17:40:00+02:00</created>
-                    dt = datetime.datetime.strptime(text[:19], "%Y-%m-%dT%H:%M:%S")
-                    setattr(self, tag, dt)
-                elif type == "date":
-                    # <date type="date">2009-10-14</date>
-                    d = datetime.date(*[int(item)for item in text.strip().split("-")])
-                    setattr(self, tag, d)
-                elif type == "float":
-                    setattr(self, tag, float(text))
-                else:
-                    if isinstance(text, str):
-                        text = text.decode("utf-8")
-                    setattr(self, tag, text)
-
-
-    def load_from_xml(self, xml_string):
-        """
-        Loads data from XML-String
-        """
-
-        # Parse XML
-        root = ET.fromstring(xml_string)
-
-        # Load
-        self.load_from_etree(root)
-
-
-    def load(self, id = None):
-        """
-        Loads the invoice-data from server
-        """
-
-        # Parameters
-        if id:
-            self.id = id
-        if not self.id:
-            raise errors.NoIdError()
-
-        # Path
-        path = "/api/invoices/{id}".format(id = self.id)
-
-        # Fetch data
-        response = self.conn.get(path = path)
-        if not response.status == 200:
-            raise errors.NotFoundError(unicode(self.id))
-
-        # Fill in data from XML
-        self.load_from_xml(response.data)
-        self.content_language = response.headers.get("content-language", None)
-
-
     def complete(self, template_id = None):
         """
         Closes a statement in the draft status (DRAFT) from.
@@ -220,7 +155,10 @@ class Invoice(Bunch):
         """
 
         # Path
-        path = "/api/invoices/{id}/email".format(id = self.id)
+        path = "{base_path}/{id}/email".format(
+            base_path = self.base_path,
+            id = self.id
+        )
 
         # XML
         email_tag = ET.Element("email")